move log parsing methods to gitscanner_log.go
